+++ extensions/year-in-progress/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Year in Progress Changelog
 
+## [Improvements] - {PR_MERGE_DATE}
+
+- Make progress icon dynamic
+
 ## [Added X In Progress] - 2023-03-29
 
 - Add X In Progress command